Changed recipe3 and added recipe5 in chapter 4.
"Recipe3 - Passing Configuration Params to Filters" as described in
the book does not match the repo's source code.

"Recipe5 - Chaining Filters together" is missing. Even though the author
explicitly says "I leave the implementation of the sortAscending filter
as an exercise to the reader...", the source code of the recipe should be
in the repo. Just like the rest of the examples.

Kept the implementation of the sortAscendingFilter as simple as possible.

The previous recipe3 is now recipe4.
The previous recipe4 is now recipe6.